src/01-mcp-servers/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
pydantic==2.11.4
fastmcp==2.3.4